+++ js/data.js
@@ -1,0 +1,31 @@
+import {getRandomInteger, getRandomIdFromRange, getRandomArrayElement} from './util';
+
+const MESSAGES = ['Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'];
+
+const NAMES = ['Андрей','Павел','Кирилл','Валерия','Карина','Ульяна','Екатерина'];
+
+const createComment = () => ({
+  id: getRandomIdFromRange(1, 6),
+  avatar: `img/avatar-${getRandomInteger(1, 6)}.svg`,
+  message: getRandomArrayElement(MESSAGES),
+  name: getRandomArrayElement(NAMES),
+});
+const createPhotoObject = () => {
+  const photoID = getRandomIdFromRange(1, 25);
+  return {
+    id: photoID,
+    url: `photos/${photoID}.jpg`,
+    description: `Описание фото номер ${photoID}`,
+    likes: getRandomInteger(15, 200),
+    comments: Array.from({length: getRandomInteger(0, 30)}, createComment),
+  };
+};
+// eslint-disable-next-line no-unused-vars
+const createPhotos = () => Array.from({length: 25}, createPhotoObject);
+
+export {createPhotos};
